requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
easyocr==1.7.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.2.0
scipy==1.11.4
scikit-image==0.22.0
streamlit==1.37.0
paddlepaddle
paddleocr
rapidocr-onnxruntime==1.3.21

